Move chunkArray() to shared
