+++ src/shared/utils/misc.ts
@@ -1,0 +1,9 @@
+export function chunkArray<T>(array: T[], chunkSize: number): T[][] {
+    const chunks: T[][] = [];
+  
+    for (let i = 0; i < array.length; i += chunkSize) {
+      chunks.push(array.slice(i, i + chunkSize));
+    }
+  
+    return chunks;
+  }
